* fixed "A LOT" of bugs * Improved error handling. All done() have a rejected function implemented now. "assertFulfilled" uses "die" now instead of "assert", since some child processes still did not die on error. This worked better.
* big update on usage of "chain" or "parallel".
"chain" before: it would already execute all promises passed in the array, however the results would be chained together
"chain" now: it will actually execute the promises one after each other. It requires an array of ToPromiseFunction now, which is a Closure returning the promise. The big difference is that has control now when the promise will be injected into the loop.
Also "parallel" works with ToPromiseFunction array now.

* Choose to execute one queued batch at a time. So "executeQueuedBatchesFor" has been changed to "executeNextQueuedBatchFor". This is because now we can send a single message to the client if that batch 'success' is either true of false like it was.
* More -Async methods added and tested, expending the possibilities of "ExecuteBatch" async.
* Simplified Router::executeCallAsync