Some bug fixes (#25)
Bug-fix 1: production was able to open up /_tools when should not be allowed

Bug-fix 2:
Problem: ReactPHP will run the loop once it created itself at the end of php script. But if we do a /1/api/Config call the AsyncDatabase is created, which creates the loop, but we do not use it , so will end into a endless loop running the end

we need to allow the AsyncDatabase to be retrieved even if async is not (yet) enabled. But it might still happen, so it is better to have this AsyncDatabase instance being created and passed along between CommonBase instances. 

Now, to prevent the endless loop, we just run a dummy awaited async task, to let ReactPHP know we have "ran" the loop, and it does not to do it and the end of script anymore